refactor: endpoints now have their APIView class in their own module inside api.views module
